requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.1
about-time==3.1.1
alive-progress==2.0.0
anyio==3.3.0
argh==0.26.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
BareNecessities==0.2.8
beautifulsoup4==4.10.0
better-profanity==0.7.0
bleach==4.1.0
blessings==1.7
blinker==1.4
bpython==0.21
brand==0.0.4
cachetools==4.2.4
caer==1.9.8
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
config2py==0.0.3
curtsies==0.3.5
cwcwidth==0.1.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
django-qrcode==0.3
dnspython==2.1.0
docopt==0.6.2
dol==0.1.41
dominate==2.6.0
email-validator==1.1.3
emoji-flag-converter==1.0
emoji-flags==0.1.0
entrypoints==0.3
fastcluster==1.2.4
filelock==3.0.12
flags==0.0.1.2
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-core==2.0.1
google-auth==2.2.1
google-auth-oauthlib==0.4.6
google-cloud-texttospeech==2.5.2
googleapis-common-protos==1.53.0
grapheme==0.6.0
greenlet==1.1.1
grpcio==1.39.0
gTTS==2.2.3
gunicorn==20.1.0
huggingface-hub==0.0.12
idna==3.2
importlib-resources==5.2.2
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.2
Mail==2.1.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mongoengine==0.23.1
mypy==0.910
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
packaging==21.0
pandas==1.3.3
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
phonenumbers==8.12.31
pickleshare==0.7.5
Pillow==8.3.2
pipreqs==0.4.10
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.19.2
protobuf==3.18.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py2store==0.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
Pygments==2.10.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
pyxdg==0.27
pyzmq==22.3.0
qrcode==7.3.1
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.9.30
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
WTForms==2.3.3
yarg==0.1.9
zipp==3.6.0
